creating about section
