--- personal/src/Styles/About.scss
@@ -1,18 +1,0 @@
-.aboutus {
-    background-image: url("../Resources/aboutplaceholder.jpg")!important;
-    width: 100%;
-    height: 600px;
-    max-height: 1000px;
-    background-repeat: no-repeat;
-    max-height: 1000px;
-    background-size: 100%;
-}
-
-.title{
-    font-size: 5rem;
-    //line-height: 90px;
-    text-align: center;
-    color: #fff;
-    padding-top: 180px;
-
-}
