+++ personal/src/Styles/variables.scss
@@ -16,7 +16,7 @@ $spacing-lg: 32px;
 $borderRadius: 12px;
 
 //header hight
-$headerHeightDesktop: 150px;
+$headerHeightDesktop: 120px;
 $headerHeightMobile: 60px;
 
 @mixin breakpoint($point){
