+++ personal/src/components/AboutSection/About.module.scss
@@ -1,0 +1,2 @@
+@import './../../Styles/variables.scss';
+
