+++ personal/src/index.js
@@ -2,8 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import "./Styles/index.scss"
+import {HashRouter} from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <App />
+    <HashRouter>
+        <App />
+    </HashRouter>
+    
 );
