--- personal/src/pages/About.js
@@ -1,58 +1,0 @@
-import React from 'react'
-import "../Styles/About.scss"
-import Header from "../components/Header"
-
-function About() {
-  return (
-      <>
-        <Header/>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        <section className="aboutus">
-            <div className="title">
-                    Welcome
-            </div>
-        </section>
-        
-  
-
-        
-      </>
-    
-  )
-}
-
-export default About
