--- personal/src/pages/Home.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import Welcome from '../components/Welcome'
-
-const Home = () => {
-  return (
-    <>
-        <section>
-            <Welcome/>
-        </section>
-    </>
-  )
-}
-
-export default Home
